lambda-layer/requirements.txt
requests==2.31.0
pandas==2.0.3
toml==0.10.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
mysql-connector-python==8.3.0
urllib3==1.25.7